+++ pages/index.js
@@ -2,6 +2,7 @@ import { Component } from 'react';
 import Info from '../components/Info';
 import Calendar from '../components/Calendar';
 import ImageCarousel from '../components/ImageCarousel';
+import { getRegisteredTeams, getRegisteredPrograms } from '../util/cms';
 import styles from '../styles/page/Index.module.css';
 import { google } from 'googleapis';
 import { createEvents } from 'ics';
@@ -607,14 +608,8 @@ export async function getStaticProps() {
     apiVersion: '2019-01-29',
     useCdn: false,
   });
-  let registeredTeams;
-  await client
-    .fetch('*[_type == "team"]{team}')
-    .then((teams) => (registeredTeams = teams));
-  let registeredPrograms;
-  await client
-    .fetch('*[_type == "program"]{program}')
-    .then((programs) => (registeredPrograms = programs));
+  const registeredTeams = await getRegisteredTeams();
+  const registeredPrograms = await getRegisteredPrograms();
   let data;
   await client
     .fetch('*[_type == "homepage"]{who, what, "images": images[].asset->url}')
