+++ pages/program/tip.js
@@ -1,5 +1,6 @@
 import { Component } from 'react';
 import ProgramPage from '../../components/ProgramPage';
+import { getRegisteredTeams, getRegisteredPrograms } from '../../util/cms';
 import { teams } from '../index';
 
 export default class TIPPage extends Component {
@@ -39,14 +40,8 @@ export async function getStaticProps() {
       '*[_type == "program" && lower(program) == "tip"]{program, accent, left, right, why, benefits, link, how, "artifacts": artifacts[]->{project, tag, contributors, description, repo}, "images": images[].asset->url, "testimonials": testimonials[]{name, description, "image": image.asset->url}}',
     )
     .then((teams) => (data = teams[0]));
-  let registeredTeams;
-  await client
-    .fetch('*[_type == "team"]{team}')
-    .then((teams) => (registeredTeams = teams));
-  let registeredPrograms;
-  await client
-    .fetch('*[_type == "program"]{program}')
-    .then((programs) => (registeredPrograms = programs));
+  const registeredTeams = await getRegisteredTeams();
+  const registeredPrograms = await getRegisteredPrograms();
   return {
     props: { data, registeredTeams, registeredPrograms },
   };
