+++ pages/scholarship.js
@@ -1,5 +1,6 @@
 import { Component } from 'react';
 import Info from '../components/Info';
+import { getRegisteredTeams, getRegisteredPrograms } from '../util/cms';
 import styles from '../styles/page/Scholarship.module.css';
 export default class ScholarshipPage extends Component {
   constructor(props) {
@@ -45,14 +46,8 @@ export async function getStaticProps() {
     apiVersion: '2019-01-29',
     useCdn: false,
   });
-  let registeredTeams;
-  await client
-    .fetch('*[_type == "team"]{team}')
-    .then((teams) => (registeredTeams = teams));
-  let registeredPrograms;
-  await client
-    .fetch('*[_type == "program"]{program}')
-    .then((programs) => (registeredPrograms = programs));
+  const registeredTeams = await getRegisteredTeams();
+  const registeredPrograms = await getRegisteredPrograms();
   let data;
   await client
     .fetch('*[_type == "scholarshippage"]{who, how, inspiration}')
