+++ pages/sponsors.js
@@ -1,4 +1,5 @@
 import { Component } from 'react';
+import { getRegisteredTeams, getRegisteredPrograms } from '../util/cms';
 import styles from '../styles/page/Sponsors.module.css';
 
 export default class SponsorsPage extends Component {
@@ -41,14 +42,8 @@ export async function getStaticProps() {
     apiVersion: '2019-01-29',
     useCdn: false,
   });
-  let registeredTeams;
-  await client
-    .fetch('*[_type == "team"]{team}')
-    .then((teams) => (registeredTeams = teams));
-  let registeredPrograms;
-  await client
-    .fetch('*[_type == "program"]{program}')
-    .then((programs) => (registeredPrograms = programs));
+  const registeredTeams = await getRegisteredTeams();
+  const registeredPrograms = await getRegisteredPrograms();
   let data;
   await client
     .fetch(
