+++ styles/globals.css
@@ -1,7 +1,3 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
 @import url('https://fonts.googleapis.com/css2?family=Poppins&display=swap');
 :root {
   --bg-color: #131313;
@@ -21,4 +17,5 @@ body {
   background-color: var(--bg-color);
   font-family: 'Poppins';
   color: white;
+  font-size: 1.5vmin;
 }
